+++ Tests/SegnifyTests/SegnifyTests.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import Segnify
+
+final class SegnifyTests: XCTestCase {
+    
+}