+++ iOS Example/Swift Package Manager/Segnified/Delegates/PageViewControllerDelegate.swift
@@ -1,0 +1,20 @@
+//
+//  PageViewControllerDelegate.swift
+//  Segnified
+//
+//  Created by Bart Hopster on 23/11/2018.
+//  Copyright © 2018 Bart Hopster. All rights reserved.
+//
+
+import UIKit
+import Segnify
+
+/// Implements `PageViewControllerProtocol`.
+public class PageViewControllerDelegate: PageViewControllerProtocol {
+    
+    // MARK: - Delegate
+    
+    public var backgroundColor: UIColor = .white
+    
+    public var segnifyHeight: CGFloat = 75.0
+}